refactor: Firebase 인증 구조 개선 #19
