--- youtube-ari/src/context/AuthContext.js
@@ -1,4 +1,0 @@
-import { createContext } from "react";
-
-// AuthContext만 내보냄.
-export const AuthContext = createContext();
